Implement switch() (SDF3 + desugar + nabl2)
This implementation is slightly more strict than desired. It allows only expressions in a switch() expression OR only statements in a switch() statement. In the future, it would be nice to also support statements in a switch() exp if those statements branch out of the place where the switch() exp's value is used (e.g. throw, break(>2), continue(), return()).