Update IVar type checking
For each typed variable declaration: Create a declaration in a new scope, add an edge to that scope from the current scope and prefer that edge over declarations in the current scope.
For each variable reference: References will automatically resolve to typed declarations, and untyped declarations when no typed declaration is available.
This effectively adds the type used to declare a variable with to all variable references.
Note that in the ideal situation, types of the assigned values should be used for variable references. This commit does not implement that.